requirements.txt
Django==5.2.12
djangorestframework==3.17.1
django-cors-headers==4.9.0
python-dotenv==1.2.2
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.4.0
